LayoutTests/http/tests/referrer-policy-script/no-referrer/same-origin.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of no-referrer referrer policy when same origin.");
jsTestIsAsync = true;

function checkReferrer(value) {
    referrer = value;
    // Should be the empty string
    shouldBeEqualToString("referrer", "");
    finishJSTest();
}
</script>
<script src="http://127.0.0.1:8000/referrer-policy/resources/script.php" referrerpolicy="no-referrer"></script>
</body>
</html>
