LayoutTests/http/tests/referrer-policy-script/no-referrer/same-origin-expected.txt

Tests the behavior of no-referrer referrer policy when same origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is ""
PASS successfullyParsed is true

TEST COMPLETE
